# Conflicts:
#	README.md
#	qtome/AppearanceSettingsWidget.cpp
#	qtome/GLView3D.cpp
#	renderlib/Histogram.cpp
#	renderlib/ImageXyzcCuda.h
#	renderlib/gl/Image2D.cpp
#	renderlib/gl/Image2D.h
#	renderlib/gl/v33/V33Image3D.cpp
#	renderlib/glsl/v330/V330GLImageShader2DnoLut.cpp

+++ renderlib/cudarndr/Variance.h
@@ -2,6 +2,8 @@
 
 #include "Defines.h"
 
+#include <math.h>
+
 class CStatus;
 
 class CVariance
@@ -47,7 +49,7 @@ public:
 
 	HOD float Mean(int ID) const
 	{
-		return (m_pN[ID] > 0) ? m_pNewM[ID] : 0.0;
+		return (m_pN[ID] > 0) ? m_pNewM[ID] : 0.0f;
 	}
 
 	HOD float GetVariance(int ID) const
@@ -57,7 +59,7 @@ public:
 
 	HOD float StandardDeviation(int ID) const
 	{
-		return sqrt( GetVariance(ID) );
+		return sqrtf( GetVariance(ID) );
 	}
 
 private:
